Use React Router v6 for the routable tabs of client scope page
